% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-test-ids.R
\name{assign_test_ids}
\alias{assign_test_ids}
\title{Assigns Test Id's to stories generated via milestones}
\usage{
assign_test_ids(test_path = getOption("TEST_DIR"), overwrite_tests = TRUE)
}
\arguments{
\item{test_path}{path to where tests are written.}

\item{overwrite_tests}{(T/F) whether or not to overwrite test files with new test ids}
}
\description{
This function will generate Test Id's with the pattern 'TST-FOO-XXX',
overwrite test documents to include these new IDs, and return a dataframe
with the new TestIds column.
}
