% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec-yaml.R
\name{check_uniq_col_vals}
\alias{check_uniq_col_vals}
\title{Check unique values of column}
\usage{
check_uniq_col_vals(df, col_uniq)
}
\arguments{
\item{df}{Tibble to check}

\item{col_uniq}{Column that should be unique}
}
\description{
Check unique values of column
}
\keyword{internal}
