% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_issue_summary}
\alias{extract_issue_summary}
\title{Extract Summary section from Github issue}
\usage{
extract_issue_summary(txt)
}
\arguments{
\item{txt}{The raw text scraped from the body of the Github issue}
}
\description{
Extract Summary section from Github issue
}
\keyword{internal}
