% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_issue_tests}
\alias{extract_issue_tests}
\title{Extract Tests section from Github issue}
\usage{
extract_issue_tests(txt)
}
\arguments{
\item{txt}{The raw text scraped from the body of the Github issue}
}
\description{
Extract Tests section from Github issue
}
