% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-github-issues.R
\name{get_risk}
\alias{get_risk}
\title{Get risk field for each issue}
\usage{
get_risk(org, repo, domain = VALID_DOMAINS)
}
\arguments{
\item{org}{Github organization that the repo is under}

\item{repo}{The name of the repo for the package you are validating}

\item{domain}{Domain where repo lives. Either "github.com" or "ghe.metrumrg.com", defaulting to "github.com"}
}
\description{
Get risk field for each issue
}
