% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tests.R
\name{get_testing_dir}
\alias{get_testing_dir}
\title{Get tests/testthat directory from package directory}
\usage{
get_testing_dir(pkg_dir)
}
\arguments{
\item{pkg_dir}{a file path pointing to an unpacked/untarred package directory}
}
\description{
Get tests/testthat directory from package directory
}
\keyword{internal}
