% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tests.R
\name{get_tests}
\alias{get_tests}
\title{Tabulate all tests in a package directory}
\usage{
get_tests(pkg_dir)
}
\arguments{
\item{pkg_dir}{a file path pointing to an unpacked/untarred package directory}
}
\value{
a dataframe containing all test names and the test file they are associated with
}
\description{
Tabulate all tests in a package directory
}
