% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-results.R
\name{git_results}
\alias{git_results}
\alias{test_and_push_results}
\alias{ingest_and_push_results}
\title{Store test results in dedicated Git reference}
\usage{
test_and_push_results(
  tests_dir,
  executor = NULL,
  info = NULL,
  remote = "origin",
  ref = "refs/mrgval/test-results",
  force = FALSE
)

ingest_and_push_results(
  results_dir,
  commit = "HEAD",
  remote = "origin",
  ref = "refs/mrgval/test-results",
  force = FALSE
)
}
\arguments{
\item{tests_dir}{Directory containing test files (e.g., "tests/testthat/").}

\item{executor}{Name of person executing to test. If not specified, the value
is taken from the \code{USER} environment variable.}

\item{info}{A named list with additional information that should be included
in the JSON output file.}

\item{remote}{Push results to this remote.}

\item{ref}{Store the test results in this Git reference and push it to
\code{remote}. This must be a fully qualified reference, and it should typically
\emph{not} be underneath the standard "refs/heads/" namespace.}

\item{force}{If test results for this commit already exist, overwrite them
rather than aborting. Note that the old results will still exist in the
history of \code{ref}.}

\item{results_dir}{Directory with results to ingest.}

\item{commit}{Commit on which the tests were executed.}
}
\description{
\code{test_and_push_results()} runs tests with \code{\link[=test_and_write_results]{test_and_write_results()}} and, if
there are no failures, stores the results in a Git reference and pushes that
to a remote.

\code{ingest_and_push_results()}, on the other hand, updates a Git reference with
the results from the specified directory and pushes that to a remote.
}
