% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec-gsheet.R
\name{gsheet_to_yaml}
\alias{gsheet_to_yaml}
\title{Read stories from a Google Sheet.}
\usage{
gsheet_to_yaml(
  ss,
  sheet = NULL,
  file = "",
  story_id_col = "StoryId",
  story_name_col = "StoryName",
  story_description_col = "StoryDescription",
  risk_col = "ProductRisk",
  test_ids_col = "TestIds"
)
}
\arguments{
\item{ss, sheet}{Sheet identifiers passed \code{\link[googlesheets4:range_read]{googlesheets4::read_sheet()}}.}

\item{file}{output}

\item{story_id_col, story_name_col, story_description_col, risk_col, test_ids_col}{Names of relevant columns in input Google Sheet.}
}
\description{
Read stories from a Google Sheet.
}
