% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-test-output.R
\name{leading_lcs}
\alias{leading_lcs}
\title{Leading longest common substring}
\usage{
leading_lcs(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Get the longest string contained in \emph{each} element of \code{x}, \emph{starting from the
beginning} of each element.
}
\keyword{internal}
