% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec-gsheet.R
\name{merge_requirements_and_stories}
\alias{merge_requirements_and_stories}
\title{Join the stories and requirements by requirement ID.}
\usage{
merge_requirements_and_stories(stories, reqs)
}
\arguments{
\item{stories}{Data frame of stories, where the RequirementIds column has a
list of requirement IDs associated with each story.}

\item{reqs}{Data frame of requirements, with each row identified by a unique
requirement ID.}
}
\description{
Join the stories and requirements by requirement ID.
}
\keyword{internal}
