% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-test-ids.R
\name{milestone_to_test_id}
\alias{milestone_to_test_id}
\title{Update stories dataframe with newly generated test ids}
\usage{
milestone_to_test_id(stories_df, tests, return_missing_ids = FALSE)
}
\arguments{
\item{stories_df}{a dataframe of stories returned by \code{parse_github_issues()}.}

\item{tests}{dataframe returned by \code{\link[=assign_test_ids]{assign_test_ids()}}
Must have the following column names: "TestNames", "TestIds"}

\item{return_missing_ids}{logical (T/F). Whether or not to return the warning messages pertaining to missing ids.
Note that this will affect piping to other functions, as the returned object will be a named list.}
}
\description{
Update stories dataframe with newly generated test ids
}
\details{
Users must first run \code{\link[=assign_test_ids]{assign_test_ids()}} to generate a list of test ids and overwrite existing test files.
}
