% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-test-ids.R
\name{overwrite_tests}
\alias{overwrite_tests}
\title{Overwrite test files with new Test Id's}
\usage{
overwrite_tests(test_scripts, TestIds)
}
\arguments{
\item{test_scripts}{list of test files}

\item{TestIds}{dataframe of test Ids, generated in \code{assign_test_ids()}}
}
\description{
Overwrite test files with new Test Id's
}
\details{
Make sure to set the following options for testing purposes:
\code{options(TEST_DIR = system.file("fake-tests", package = "mrgvalprep"))}
\code{options(TEST_DIR_TESTING = TRUE)}
}
\keyword{internal}
