% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-test-output.R
\name{parse_golang_test_json}
\alias{parse_golang_test_json}
\title{Parse Go test output}
\usage{
parse_golang_test_json(test_file)
}
\arguments{
\item{test_file}{Path to a file \code{.json} file containing the output from \verb{go test --json}}
}
\value{
A tibble formatted according to \code{mrgvalidate::input_formats}
}
\description{
Parses output of \verb{go test --json} into a tibble for \code{mrgvalidate} to consume.
}
\details{
\itemize{
\item Any tests \emph{without} Test Id's (correctly formatted \emph{in brackets}) will be thrown out.
\item All tests and subtests with the same Test Id will be rolled up to a single
row in the output tibble. The \verb{passed,failed} counts will reflect the number
of tests/subtests that were rolled up this way.
\item Any test with \emph{more than one Test Id} will have only the first Test Id
extracted. Any subsequent Test Id's for that test will be ignored. Generally,
the first Id will be the \emph{least} specific, since subsequent Id's will likely
have been added as part of subtests. Keep this in mind when adding Testd Id's
in your Go test code.
}
}
\seealso{
\code{mrgvalidate::input_formats}, \code{mrgvalidate::create_validation_docs()}
}
