% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-test-ids.R
\name{parse_tests_from_files}
\alias{parse_tests_from_files}
\title{Reads and returns all tests from test files}
\usage{
parse_tests_from_files(test_scripts, test_type)
}
\arguments{
\item{test_scripts}{character vector. File paths of tests}

\item{test_type}{string. Denotes whether the tests were written via \code{test_that()} or
\code{describe()}/\code{it()} functionality. Specify as 'test_that' or 'it'}
}
\description{
Reads and returns all tests from test files
}
\keyword{internal}
