% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proc_issue}
\alias{proc_issue}
\title{Process a single issue}
\usage{
proc_issue(txt)
}
\arguments{
\item{txt}{The raw text scraped from the body of the Github issue}
}
\description{
Process a single issue
}
