% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-stories.R
\name{process_stories}
\alias{process_stories}
\title{Process issues from Github into a tibble of properly formatted stories}
\usage{
process_stories(
  release_issues,
  org,
  repo,
  domain = VALID_DOMAINS,
  test_path = ALL_TESTS,
  add_risk = TRUE
)
}
\arguments{
\item{release_issues}{Tibble output from \code{\link[=get_issues]{get_issues()}} or \code{\link[ghpm:get_issues]{ghpm::get_issues()}} containing all the issues you want to format.}

\item{org}{Github organization that the repo is under}

\item{repo}{The name of the repo for the package you are validating}

\item{domain}{Domain where repo lives. Either "github.com" or "ghe.metrumrg.com", defaulting to "github.com"}

\item{test_path}{File path to read test information from. Defaults to "all_tests.csv". If NULL is passed, no test information will be included.}

\item{add_risk}{Boolean for whether to pull the risk field from issues as well. Defaults to TRUE.}
}
\description{
Process issues from Github into a tibble of properly formatted stories
}
