% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-tests.R
\name{pull_tagged_repo}
\alias{pull_tagged_repo}
\title{Clones the specified repo at the specified tag to disk}
\usage{
pull_tagged_repo(
  org,
  repo,
  tag,
  domain = VALID_DOMAINS,
  dest_dir = tempdir(),
  overwrite = TRUE
)
}
\arguments{
\item{org}{Github organization that the repo is under}

\item{repo}{The name of the repo for the package you are validating}

\item{tag}{The tag to pull from the repo. When this function is called internally, this is assumed to be the same as the version you are testing, though it can be any valid tag.}

\item{domain}{Domain where repo lives. Either "github.com" or "ghe.metrumrg.com", defaulting to "github.com"}

\item{dest_dir}{File path for directory to clone repo into. Defaults to \code{\link[=tempdir]{tempdir()}}}

\item{overwrite}{Boolean indicating whether to overwrite \code{file.path(dest_dir, repo)} if something already exists there. TRUE by default.}
}
\description{
Clones the specified repo at the specified tag to disk
}
\keyword{internal}
