% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-specs.R
\name{read_stories_gsheet}
\alias{read_stories_gsheet}
\title{Read stories from a Google Sheet.}
\usage{
read_stories_gsheet(
  ss,
  sheet = NULL,
  story_id_col = "StoryId",
  story_name_col = "StoryName",
  story_description_col = "StoryDescription",
  risk_col = "ProductRisk",
  req_ids_col = "RequirementIds"
)
}
\arguments{
\item{ss, sheet}{Sheet identifiers passed \code{\link[googlesheets4:range_read]{googlesheets4::read_sheet()}}.}

\item{story_id_col, story_name_col, story_description_col, risk_col, req_ids_col}{Names of relevant columns in input Google Sheet.}
}
\value{
Tibble with the above columns.
}
\description{
Read stories from a Google Sheet.
}
\keyword{internal}
