% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec-gsheet.R
\name{read_stories_only_gsheet}
\alias{read_stories_only_gsheet}
\title{Read a stories gsheet that maps to tests instead of requirements}
\usage{
read_stories_only_gsheet(
  ss,
  sheet = NULL,
  story_id_col = "StoryId",
  story_name_col = "StoryName",
  story_description_col = "StoryDescription",
  risk_col = "ProductRisk",
  test_ids_col = "TestIds"
)
}
\description{
Read a stories gsheet that maps to tests instead of requirements
}
\keyword{internal}
