% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_test_df}
\alias{read_test_df}
\title{Read in the csv will results from all the tests}
\usage{
read_test_df(test_path = ALL_TESTS)
}
\arguments{
\item{test_path}{File path to file containing test results}
}
\description{
Read in the csv will results from all the tests
}
