% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-test-ids.R
\name{replace_test_str}
\alias{replace_test_str}
\title{Replace test names in test file}
\usage{
replace_test_str(test_file, from, to)
}
\arguments{
\item{test_file}{character vector of test file}

\item{from}{character vector. To replace}

\item{to}{character vector. Replacement}
}
\description{
Replace test names ('str') in test file with string containing new TestId ('str' + \code{TST-FOO-XXX})
}
\keyword{internal}
