% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirements-from-tests.R
\name{requirements_from_tests}
\alias{requirements_from_tests}
\alias{req_df_from_tests}
\alias{req_df_to_yaml}
\alias{stories_replace_tests_with_reqs}
\title{Requirements from tests}
\usage{
req_df_from_tests(test_df)

req_df_to_yaml(req_df, out_file)

stories_replace_tests_with_reqs(stories_file, req_spec)
}
\description{
If you were previously linking stories directly to tests, and
you would like to begin using requirements, use \code{\link[=req_df_from_tests]{req_df_from_tests()}} and
\code{\link[=req_df_to_yaml]{req_df_to_yaml()}} to create a requirements YAML file that parses
requirement descriptions from test names. Then use
\code{\link[=stories_replace_tests_with_reqs]{stories_replace_tests_with_reqs()}} to edit your stories YAML to reference
the newly created requirements.
}
\details{
Note that these functions are designed to help you convert a repo
that previously did \emph{not} use requirements into one that does. They are not
meant for continued use after that. Going forward, requirements should be
specified in the requirements YAML file, instead of being pull through from
test names
}
\section{Functions}{
\itemize{
\item \code{req_df_from_tests}: Create new requirements tibble from test
results tibble.

\item \code{req_df_to_yaml}: Write requirements YAML file from
requirements tibble.

\item \code{stories_replace_tests_with_reqs}: Modify existing stories YAML to reference
new requirements instead of tests.
}}

\examples{
\dontrun{

# run your test suite to get a tibble of test names and IDs
test_res <- mrgvalprep::parse_testthat_list_reporter(
  devtools::test(Reporter = testthat::ListReporter)
)

# create requirements tibble and write to new YAML
req_df <- req_df_from_tests(test_res)
req_df_to_yaml(req_df, "inst/validation/stories.yaml")

# edit existing stories YAML
stories_replace_tests_with_reqs("inst/validation/stories.yaml", req_df)

}
}
