% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-test-output.R
\name{roll_up_test_ids}
\alias{roll_up_test_ids}
\title{Roll up so test ID's are unique and passed/failed sums up counts}
\usage{
roll_up_test_ids(test_df)
}
\description{
Roll up so test ID's are unique and passed/failed sums up counts
}
\keyword{internal}
