% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-tests.R
\name{run_tests}
\alias{run_tests}
\title{Test a source package}
\usage{
run_tests(pkg, test_path = "tests/testthat", root_dir = tempdir())
}
\arguments{
\item{pkg}{Name of the package to test; should be installed to a library in
\code{\link[=.libPaths]{.libPaths()}}.}

\item{test_path}{Directory containing tests, where \code{\link[testthat:test_dir]{testthat::test_dir()}}
will be run.}

\item{root_dir}{The directory path to where the package is (i.e. where the
repo has been cloned). \code{file.path(root_dir, pkg, test_path)} should lead to
the directory that will be tested.}
}
\description{
This is an internal function called by \code{\link[=validate_tests]{validate_tests()}}
}
\keyword{internal}
