% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-tests.R
\name{setup_package_env}
\alias{setup_package_env}
\title{Helper for setting up package environment for testing}
\usage{
setup_package_env(package, test_path)
}
\arguments{
\item{package}{the package name}

\item{test_path}{path to folder with tests in it}
}
\description{
These are all things that were copied out of internal functions called by testthat::test_check to set up the package environment.
They are necessary because testthat::test_dir does NOT do this, which causes some tests to fail.
Specifically, this code came mostly from testthat:::test_package_dir and testthat:::test_pkg_env and some from test_check itself.
}
