% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-test-ids.R
\name{sort_tests_by_nchar}
\alias{sort_tests_by_nchar}
\title{Sort test ids by number of characters in test description}
\usage{
sort_tests_by_nchar(test_ids)
}
\arguments{
\item{test_ids}{data.frame of test ids}
}
\description{
Sort test ids by number of characters in test description
}
\keyword{internal}
