% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-test-output.R
\name{strip_test_id}
\alias{strip_test_id}
\title{Return a string without the embedded test ID.}
\usage{
strip_test_id(string, id)
}
\description{
Note that any white space in the string will be collapsed to a single
character.
}
