% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-results.R
\name{test_and_write_results}
\alias{test_and_write_results}
\title{Run tests and write output files for mrgvalidate}
\usage{
test_and_write_results(tests_dir, results_dir, executor = NULL, info = NULL)
}
\arguments{
\item{tests_dir}{Directory containing test files (e.g., "tests/testthat/").}

\item{results_dir}{Write the results to this directory. If this exists, it
must be empty.}

\item{executor}{Name of person executing to test. If not specified, the value
is taken from the \code{USER} environment variable.}

\item{info}{A named list with additional information that should be included
in the JSON output file.}
}
\description{
mrgvalidate requires test results to be written in a particular format (see
\code{mrgvalidate::input_formats}). The function runs the tests underneath a
directory, writing the results to a specified directory.
}
\details{
To use this function, the tests must be a part of a package and be underneath
the current directory's Git repository.
}
\seealso{
\code{\link[=test_and_push_results]{test_and_push_results()}}, \code{\link[=parse_testthat_list_reporter]{parse_testthat_list_reporter()}},
\code{\link[=get_sys_info]{get_sys_info()}}
}
