% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-tests.R
\name{validate_tests}
\alias{validate_tests}
\title{Run and summarize tests}
\usage{
validate_tests(
  pkg,
  root_dir = tempdir(),
  out_file = ALL_TESTS,
  output_dir = getwd(),
  return_df = FALSE,
  extra_test_dirs = NULL
)
}
\arguments{
\item{pkg}{The name of the package you are validating, to be included in the
output document.}

\item{root_dir}{The directory path to where the package has been cloned.
\code{file.path(root_dir, pkg)} should lead to the cloned repo that will be
tested.}

\item{out_file}{File path to write out the test results to. Any extension
will be ignored and replaced with .csv}

\item{output_dir}{Directory to write the output documents to. Defaults to
working directory.}

\item{return_df}{Boolean indicating whether to return the tibble that is
written to \code{out_file}. Defaults to FALSE and returns nothing.}

\item{extra_test_dirs}{Character vector of paths (relative to package root
dir) to directories that contain additional tests to run}
}
\description{
\code{pkg} will be built and installed (along with any missing dependencies) to a
temporary library. Then we run all tests for \code{pkg}, roll up successes and
failures, and write to a CSV file.
}
\details{
The package found at \code{file.path(root_dir, pkg)} will be built and installed
into a tempory library that is appended to the \emph{beginning} of your
\code{.libPaths()}, using \code{devtools::install(..., build = TRUE, upgrade = "never")}. This will also install (into the temporary library) any
dependencies from the package that are \emph{not} found in your \code{.libPaths()}.
}
