% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-traceability-matrix.R
\name{write_traceability_matrix}
\alias{write_traceability_matrix}
\title{Build the Traceability Matrix and write it to a output files}
\usage{
write_traceability_matrix(
  df,
  pkg,
  version,
  out_file = MAT_FILE,
  output_dir = getwd(),
  word_document = TRUE
)
}
\arguments{
\item{df}{Tibble output from \code{\link[=process_stories]{process_stories()}}.}

\item{pkg}{The name of the package you are validating, to be included in the output document.}

\item{version}{The version number of the package you are validating, to be included in the output document.}

\item{out_file}{filename to write markdown file out to. Any extension will be ignored and replaced with .md}

\item{output_dir}{Directory to write the output documents to. Defaults to working directory.}

\item{word_document}{Logical scaler indicating whether to render a docx document}
}
\description{
Build the Traceability Matrix and write it to a output files
}
