% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-validation-testing.R
\name{write_validation_testing}
\alias{write_validation_testing}
\title{Build the Validation Testing document and write it to a markdown file}
\usage{
write_validation_testing(
  org,
  repo,
  version,
  domain = VALID_DOMAINS,
  extra_test_dirs = NULL,
  out_file = VAL_FILE,
  output_dir = getwd(),
  word_document = TRUE,
  dry_run = FALSE
)
}
\arguments{
\item{org}{Github organization that the repo is under}

\item{repo}{The name of the repo for the package you are validating}

\item{version}{The version number of the package you are validating. This must correspond to a tag in the repo.}

\item{domain}{Domain where repo lives. Either "github.com" or "ghe.metrumrg.com", defaulting to "github.com"}

\item{extra_test_dirs}{Character vector of paths (relative to package root dir) to directories that contain additional tests to run}

\item{out_file}{filename to write markdown file out to. Any extension will be ignored and replaced with .md}

\item{output_dir}{Directory to write the output documents to. Defaults to working directory.}

\item{word_document}{Logical scaler indicating whether to render a docx document}

\item{dry_run}{Boolean indicating whether to clone repo and run tests (if FALSE) or skip those steps and look for test results in "all_tests.csv" instead (if TRUE). FALSE by default.}
}
\description{
This function will pull
}
