% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec-gsheet.R
\name{gsheet_to_yaml}
\alias{gsheet_to_yaml}
\title{Read stories from a Google Sheet and write to a yaml file.}
\usage{
gsheet_to_yaml(ss, sheet = NULL, file)
}
\arguments{
\item{ss, sheet}{Sheet identifiers passed \code{\link[googlesheets4:range_read]{googlesheets4::read_sheet()}}.}

\item{file}{output}

\item{story_id_col, story_name_col, story_description_col, risk_col, test_ids_col}{Names of relevant columns in input Google Sheet.}
}
\description{
Read stories from a Google Sheet and write to a yaml file.
}
\details{
\subsection{Option 1: Stories and Requirements}{

(Note: Option 1 is not currently implemented for \code{gsheet_to_yaml()})
}

\subsection{Option 2: Stories linked directly to Tests}{

The stories sheet passed to \code{ss_stories} must have the following columns:
\itemize{
\item \code{StoryId} (character scalar)
\item \code{StoryName} (character scalar)
\item \code{StoryDescription} (character scalar)
\item \code{ProductRisk} (character scalar)
\item \code{TestIds} (character vector)
}
}
}
