% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-test-ids.R
\name{milestone_to_test_id}
\alias{milestone_to_test_id}
\title{Update stories dataframe with newly generated test ids}
\usage{
milestone_to_test_id(stories_df, test_ids)
}
\arguments{
\item{stories_df}{a dataframe of stories returned by \code{parse_github_issues()}.}
}
\description{
Update stories dataframe with newly generated test ids
}
\details{
Users must first run \code{\link[=assign_test_ids]{assign_test_ids()}} to generate a list of test ids and overwrite existing test files.
}
