% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-test-ids.R
\name{parse_tests}
\alias{parse_tests}
\title{Reads and returns all tests from test files}
\usage{
parse_tests(test_file)
}
\arguments{
\item{test_file}{file path to test script}
}
\description{
Reads and returns all tests from test files
}
\keyword{internal}
