% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-test-output.R
\name{parse_testthat_list_reporter}
\alias{parse_testthat_list_reporter}
\title{Parse testthat output}
\usage{
parse_testthat_list_reporter(result)
}
\arguments{
\item{result}{List output as reported by \link[testthat:ListReporter]{testthat::ListReporter}}
}
\value{
A tibble formatted according to \code{mrgvalidate::input_formats}
}
\description{
Parse testthat output
}
\seealso{
\code{mrgvalidate::input_formats}, \code{mrgvalidate::create_validation_docs()}
}
