% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-test-output.R
\name{parse_testthat_list_reporter}
\alias{parse_testthat_list_reporter}
\title{Parse testthat output}
\usage{
parse_testthat_list_reporter(result, roll_up_ids = FALSE)
}
\arguments{
\item{result}{List output as reported by \link[testthat:ListReporter]{testthat::ListReporter}}

\item{roll_up_ids}{If \code{FALSE}, the default, will leave duplicated Test ID's
as is. If \code{TRUE}, will roll up any duplicated (non-NA) ID's so that they
are unique and the passed/failed count reflects the total sums for a tests
with a given ID. The test names for a set of rolled up ID's must share a
common prefix, which will be used as the new test name. Any trailing
spaces, slashes, or underscores will be removed from the extracted prefix.}
}
\value{
A tibble formatted according to \code{mrgvalidate::input_formats}
}
\description{
Parse testthat output
}
\seealso{
\code{mrgvalidate::input_formats}, \code{mrgvalidate::create_validation_docs()}
}
