% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-specs.R
\name{read_spec_gsheets}
\alias{read_spec_gsheets}
\title{Read requirements and stories from Google Sheets.}
\usage{
read_spec_gsheets(ss_stories, ss_req, sheet_stories = NULL, sheet_req = NULL)
}
\arguments{
\item{ss_stories, ss_req, sheet_stories, sheet_req}{Sheet identifiers for the
stories and requirements passed along as the \code{ss} and \code{sheet} arguments to
\code{\link[googlesheets4:range_read]{googlesheets4::read_sheet()}}.}
}
\value{
Tibble joining requirements and stories by RequirementId.
}
\description{
Read requirements and stories from Google Sheets.
}
\details{
The stories sheet must have the following columns:
\itemize{
\item StoryId (character scalar)
\item StoryName (character scalar)
\item StoryDescription (character scalar)
\item ProductRisk (character scalar)
\item RequirementIds (character vector)
}

The requirements sheet must have the following columns:
\itemize{
\item RequirementId (character scalar)
\item RequirementDescription (character scalar)
\item TestIds (character vector)
}
}
