% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec-gsheet.R
\name{read_spec_gsheets}
\alias{read_spec_gsheets}
\title{Read requirements and stories from Google Sheets.}
\usage{
read_spec_gsheets(
  ss_stories,
  ss_req = NULL,
  sheet_stories = NULL,
  sheet_req = NULL
)
}
\arguments{
\item{ss_stories, ss_req, sheet_stories, sheet_req}{Sheet identifiers for the
stories and requirements passed along as the \code{ss} and \code{sheet} arguments to
\code{\link[googlesheets4:range_read]{googlesheets4::read_sheet()}}.}
}
\value{
Tibble joining requirements and stories by RequirementId.
}
\description{
Read requirements and stories from Google Sheets.
}
\details{
\subsection{Option 1: Stories and Requirements}{

The stories sheet passed to \code{ss_stories} must have the following columns:
\itemize{
\item StoryId (character scalar)
\item StoryName (character scalar)
\item StoryDescription (character scalar)
\item ProductRisk (character scalar)
\item RequirementIds (character vector)
}

The requirements sheet passed to \code{ss_req} must have the following columns:
\itemize{
\item RequirementId (character scalar)
\item RequirementDescription (character scalar)
\item TestIds (character vector)
}
}

\subsection{Option 2: Stories linked directly to Tests}{

Don't pass anything to \code{ss_req}. The stories sheet passed to \code{ss_stories} must have the following columns:
\itemize{
\item StoryId (character scalar)
\item StoryName (character scalar)
\item StoryDescription (character scalar)
\item ProductRisk (character scalar)
\item TestIds (character vector)
}

(Note: these are the same columns as Option 1, but replacing \code{RequirementIds} with \code{TestIds})
}
}
