% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spec-gsheet.R
\name{stories_to_yaml}
\alias{stories_to_yaml}
\title{Read stories from a Google Sheet and write to a yaml file.}
\usage{
stories_to_yaml(sheet_df, file)
}
\arguments{
\item{sheet_df}{a dataframe of stories returned by \code{read_stories_only_gsheet()} or \code{parse_github_issues()}.}

\item{file}{output}
}
\description{
Read stories from a Google Sheet and write to a yaml file.
}
\details{
\subsection{Option 1: Stories and Requirements}{

(Note: Option 1 is not currently implemented for \code{gsheet_to_yaml()})
}

\subsection{Option 2: Stories linked directly to Tests}{

The stories sheet passed to \code{ss_stories} must have the following columns:
\itemize{
\item \code{StoryId} (character scalar)
\item \code{StoryName} (character scalar)
\item \code{StoryDescription} (character scalar)
\item \code{ProductRisk} (character scalar)
\item \code{TestIds} (character vector)
}
}
}
\examples{
# github milestones

# Note: Make sure to set the environment variable `GITHUB_PAT` with your personal
#       access token to authorize GHPM before using `parse_github_issues()`

MILESTONES <- c("v0.6.0", "v0.6.1")
parse_github_issues(org = "metrumresearchgroup", repo = "mrgvalidatetestreference",
 mile = MILESTONES, domain = "github.com") \%>\%
  stories_to_yaml(file = file.path(tempdir(), "temp.yaml")) \%>\%
  file.edit()


# google sheets

read_stories_only_gsheet(ss = "1LpSX5Rb1XM5-xmQ8Wl2gQjMT5-3FIkuCM7oZhSgvWeI") \%>\%
  stories_to_yaml(file = file.path(tempdir(), "temp.yaml")) \%>\%
  file.edit()

}
